improving rns prices and account management
