+++ app/upgrades/recover.go
@@ -3,6 +3,7 @@ package upgrades
 import (
 	"encoding/hex"
 	"fmt"
+	allTypes "github.com/jackalLabs/canine-chain/v4/types"
 	"strconv"
 	"strings"
 
@@ -14,7 +15,7 @@ import (
 const AccountName = "STORAGE_ESCROW"
 
 func RecoverFiles(ctx sdk.Context, keeper *storageKeeper.Keeper, merkles string, planHeight int64, name string) error {
-	account, err := types.GetAccount(AccountName) // creating account to hold the files
+	account, err := allTypes.GetAccount(AccountName) // creating account to hold the files
 	if err != nil {
 		return err
 	}
