+++ cmd/canined/utils.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
+	allTypes "github.com/jackalLabs/canine-chain/v4/types"
 
-	"github.com/jackalLabs/canine-chain/v4/x/storage/types"
 	"github.com/spf13/cobra"
 )
 
@@ -13,7 +13,7 @@ func AddressGenerationCommand() *cobra.Command {
 		Short: "Generate an address from text, the same system used to generate protocol owned accounts.",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(_ *cobra.Command, args []string) error {
-			acc, err := types.GetAccount(args[0])
+			acc, err := allTypes.GetAccount(args[0])
 			if err != nil {
 				return err
 			}
