+++ x/rns/types/tlds.go
@@ -4,4 +4,4 @@ var SupportedTLDs = []string{"ibc", "jkl"}
 
 var IsReserved = map[string]bool{"ibc": false, "jkl": false}
 
-var TLDCost = map[string]int64{"ibc": 50000000, "jkl": 25000000}
+var TLDCost = map[string]int64{"ibc": 50_000_000, "jkl": 10_000_000}
