+++ x/storage/types/keys.go
@@ -1,12 +1,10 @@
 package types
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
+	"github.com/jackalLabs/canine-chain/v4/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
 const (
@@ -28,8 +26,7 @@ const (
 	AddressPrefix = "jkl"
 	CidPrefix     = "jklc"
 
-	CollateralCollectorName    = "storage_collateral_name"
-	ProtocolOwnedLiquidityName = "protocol_owned_liq"
+	CollateralCollectorName = "storage_collateral_name"
 )
 
 func gaugeName(gauge PaymentGauge) string {
@@ -37,23 +34,7 @@ func gaugeName(gauge PaymentGauge) string {
 }
 
 func GetGaugeAccount(gauge PaymentGauge) (sdk.AccAddress, error) {
-	return GetAccount(gaugeName(gauge))
-}
-
-func GetPOLAccount() (sdk.AccAddress, error) {
-	return GetAccount(ProtocolOwnedLiquidityName)
-}
-
-func GetAccount(name string) (sdk.AccAddress, error) {
-	s := sha256.New()
-	s.Write([]byte(name))
-	m := s.Sum(nil)
-	mh := hex.EncodeToString(m)
-	adr, err := sdk.AccAddressFromHex(mh)
-	if err != nil {
-		return nil, sdkerrors.Wrapf(err, "cannot get account account")
-	}
-	return adr, nil
+	return types.GetAccount(gaugeName(gauge))
 }
 
 func KeyPrefix(p string) []byte {
